(update) to latest version

+++ plugins/admin/classes/gpm.php
@@ -2,6 +2,7 @@
 
 namespace Grav\Plugin\Admin;
 
+use Grav\Common\Cache;
 use Grav\Common\Grav;
 use Grav\Common\GPM\GPM as GravGPM;
 use Grav\Common\GPM\Licenses;
@@ -10,7 +11,6 @@ use Grav\Common\GPM\Response;
 use Grav\Common\GPM\Upgrader;
 use Grav\Common\Filesystem\Folder;
 use Grav\Common\GPM\Common\Package;
-use Grav\Plugin\Admin\Admin;
 
 /**
  * Class Gpm
@@ -50,15 +50,14 @@ class Gpm
      * @param Package[]|string[]|string $packages
      * @param array                     $options
      *
-     * @return bool
+     * @return string|bool
      */
     public static function install($packages, array $options)
     {
         $options = array_merge(self::$options, $options);
 
-        if (
-            !Installer::isGravInstance($options['destination'])
-            || !Installer::isValidDestination($options['destination'], [Installer::EXISTS, Installer::IS_LINK])
+        if (!Installer::isGravInstance($options['destination']) || !Installer::isValidDestination($options['destination'],
+                [Installer::EXISTS, Installer::IS_LINK])
         ) {
             return false;
         }
@@ -88,11 +87,11 @@ class Gpm
             // Check destination
             Installer::isValidDestination($options['destination'] . DS . $package->install_path);
 
-            if (Installer::lastErrorCode() === Installer::EXISTS && !$options['overwrite']) {
+            if (!$options['overwrite'] && Installer::lastErrorCode() === Installer::EXISTS) {
                 return false;
             }
 
-            if (Installer::lastErrorCode() === Installer::IS_LINK && !$options['ignore_symlinks']) {
+            if (!$options['ignore_symlinks'] && Installer::lastErrorCode() === Installer::IS_LINK) {
                 return false;
             }
 
@@ -109,13 +108,13 @@ class Gpm
                 throw new \RuntimeException($msg);
             }
 
-            if (count($packages) == 1) {
+            if (count($packages) === 1) {
                 $message = Installer::getMessage();
                 if ($message) {
                     return $message;
-                } else {
-                    $messages .= $message;
                 }
+
+                $messages .= $message;
             }
         }
 
@@ -126,7 +125,7 @@ class Gpm
      * @param Package[]|string[]|string $packages
      * @param array                     $options
      *
-     * @return bool
+     * @return string|bool
      */
     public static function update($packages, array $options)
     {
@@ -139,13 +138,13 @@ class Gpm
      * @param Package[]|string[]|string $packages
      * @param array                     $options
      *
-     * @return bool
+     * @return string|bool
      */
     public static function uninstall($packages, array $options)
     {
         $options = array_merge(self::$options, $options);
 
-        $packages = is_array($packages) ? $packages : [$packages];
+        $packages = (array)$packages;
         $count    = count($packages);
 
         $packages = array_filter(array_map(function ($p) {
@@ -171,7 +170,7 @@ class Gpm
             // Check destination
             Installer::isValidDestination($location);
 
-            if (Installer::lastErrorCode() === Installer::IS_LINK && !$options['ignore_symlinks']) {
+            if (!$options['ignore_symlinks'] && Installer::lastErrorCode() === Installer::IS_LINK) {
                 return false;
             }
 
@@ -183,13 +182,101 @@ class Gpm
                 throw new \RuntimeException($msg);
             }
 
-            if (count($packages) == 1) {
+            if (count($packages) === 1) {
                 $message = Installer::getMessage();
                 if ($message) {
                     return $message;
                 }
             }
         }
+
+        return true;
+    }
+
+    /**
+     * Direct install a file
+     *
+     * @param string $package_file
+     *
+     * @return string|bool
+     */
+    public static function directInstall($package_file)
+    {
+        if (!$package_file) {
+            return Admin::translate('PLUGIN_ADMIN.NO_PACKAGE_NAME');
+        }
+
+        $tmp_dir = Grav::instance()['locator']->findResource('tmp://', true, true);
+        $tmp_zip = $tmp_dir . '/Grav-' . uniqid('', false);
+
+        if (Response::isRemote($package_file)) {
+            $zip = GravGPM::downloadPackage($package_file, $tmp_zip);
+        } else {
+            $zip = GravGPM::copyPackage($package_file, $tmp_zip);
+        }
+
+        if (file_exists($zip)) {
+            $tmp_source = $tmp_dir . '/Grav-' . uniqid('', false);
+            $extracted  = Installer::unZip($zip, $tmp_source);
+
+            if (!$extracted) {
+                Folder::delete($tmp_source);
+                Folder::delete($tmp_zip);
+                return Admin::translate('PLUGIN_ADMIN.PACKAGE_EXTRACTION_FAILED');
+            }
+
+            $type = GravGPM::getPackageType($extracted);
+
+            if (!$type) {
+                Folder::delete($tmp_source);
+                Folder::delete($tmp_zip);
+                return Admin::translate('PLUGIN_ADMIN.NOT_VALID_GRAV_PACKAGE');
+            }
+
+            if ($type === 'grav') {
+                Installer::isValidDestination(GRAV_ROOT . '/system');
+                if (Installer::IS_LINK === Installer::lastErrorCode()) {
+                    Folder::delete($tmp_source);
+                    Folder::delete($tmp_zip);
+                    return Admin::translate('PLUGIN_ADMIN.CANNOT_OVERWRITE_SYMLINKS');
+                }
+
+                static::upgradeGrav($zip, $extracted);
+            } else {
+                $name = GravGPM::getPackageName($extracted);
+
+                if (!$name) {
+                    Folder::delete($tmp_source);
+                    Folder::delete($tmp_zip);
+                    return Admin::translate('PLUGIN_ADMIN.NAME_COULD_NOT_BE_DETERMINED');
+                }
+
+                $install_path = GravGPM::getInstallPath($type, $name);
+                $is_update    = file_exists($install_path);
+
+                Installer::isValidDestination(GRAV_ROOT . DS . $install_path);
+                if (Installer::lastErrorCode() === Installer::IS_LINK) {
+                    Folder::delete($tmp_source);
+                    Folder::delete($tmp_zip);
+                    return Admin::translate('PLUGIN_ADMIN.CANNOT_OVERWRITE_SYMLINKS');
+                }
+
+                Installer::install($zip, GRAV_ROOT,
+                    ['install_path' => $install_path, 'theme' => $type === 'theme', 'is_update' => $is_update],
+                    $extracted);
+            }
+
+            Folder::delete($tmp_source);
+
+            if (Installer::lastErrorCode()) {
+                return Installer::lastErrorMsg();
+            }
+
+        } else {
+            return Admin::translate('PLUGIN_ADMIN.ZIP_PACKAGE_NOT_FOUND');
+        }
+
+        Folder::delete($tmp_zip);
 
         return true;
     }
@@ -204,14 +291,11 @@ class Gpm
         $query = '';
 
         if ($package->premium) {
-            $query = \json_encode(array_merge(
-                $package->premium,
-                [
-                    'slug'        => $package->slug,
-                    'filename'    => $package->premium['filename'],
-                    'license_key' => $license
-                ]
-            ));
+            $query = \json_encode(array_merge($package->premium, [
+                'slug'        => $package->slug,
+                'filename'    => $package->premium['filename'],
+                'license_key' => $license
+            ]));
 
             $query = '?d=' . base64_encode($query);
         }
@@ -222,14 +306,13 @@ class Gpm
             throw new \RuntimeException($e->getMessage());
         }
 
-        $tmp_dir = Admin::getTempDir() . '/Grav-' . uniqid();
+        $tmp_dir = Admin::getTempDir() . '/Grav-' . uniqid('', false);
         Folder::mkdir($tmp_dir);
 
-        $bad_chars = array_merge(
-            array_map('chr', range(0, 31)),
-            array("<", ">", ":", '"', "/", "\\", "|", "?", "*"));
+        $bad_chars = array_merge(array_map('chr', range(0, 31)), ['<', '>', ':', '"', '/', '\\', '|', '?', '*']);
 
-        $filename = $package->slug . str_replace($bad_chars, "", basename($package->zipball_url));
+        $filename = $package->slug . str_replace($bad_chars, '', basename($package->zipball_url));
+        $filename = preg_replace('/[\\\\\/:"*?&<>|]+/m', '-', $filename);
 
         file_put_contents($tmp_dir . DS . $filename . '.zip', $contents);
 
@@ -268,13 +351,14 @@ class Gpm
             return false;
         }
 
-        if (method_exists($upgrader, 'meetsRequirements') && !$upgrader->meetsRequirements()) {
+        if (method_exists($upgrader, 'meetsRequirements') &&
+            method_exists($upgrader, 'minPHPVersion') &&
+            !$upgrader->meetsRequirements()) {
             $error   = [];
             $error[] = '<p>Grav has increased the minimum PHP requirement.<br />';
-            $error[] = 'You are currently running PHP <strong>' . PHP_VERSION . '</strong>';
-            $error[] = ', but PHP <strong>' . GRAV_PHP_MIN . '</strong> is required.</p>';
-            $error[] =
-                '<p><a href="http://getgrav.org/blog/changing-php-requirements-to-5.5" class="button button-small secondary">Additional information</a></p>';
+            $error[] = 'You are currently running PHP <strong>' . phpversion() . '</strong>';
+            $error[] = ', but PHP <strong>' . $upgrader->minPHPVersion() . '</strong> is required.</p>';
+            $error[] = '<p><a href="http://getgrav.org/blog/changing-php-requirements-to-5.5" class="button button-small secondary">Additional information</a></p>';
 
             Installer::setError(implode("\n", $error));
 
@@ -282,20 +366,65 @@ class Gpm
         }
 
         $update = $upgrader->getAssets()['grav-update'];
-        $tmp    = Admin::getTempDir() . '/Grav-' . uniqid();
-        $file   = self::_downloadSelfupgrade($update, $tmp);
+        $tmp    = Admin::getTempDir() . '/Grav-' . uniqid('', false);
+        if ($tmp) {
+            $file   = self::_downloadSelfupgrade($update, $tmp);
+            $folder = Installer::unZip($file, $tmp . '/zip');
+            $keepFolder = false;
+        } else {
+            // If you make $tmp empty, you can install your local copy of Grav (for testing purposes only).
+            $file = 'grav.zip';
+            $folder = '~/phpstorm/grav-clones/grav';
+            //$folder = '/home/matias/phpstorm/rockettheme/grav-devtools/grav-clones/grav';
+            $keepFolder = true;
+        }
 
-        Installer::install($file, GRAV_ROOT,
-            ['sophisticated' => true, 'overwrite' => true, 'ignore_symlinks' => true]);
+        static::upgradeGrav($file, $folder, $keepFolder);
 
         $errorCode = Installer::lastErrorCode();
 
-        Folder::delete($tmp);
-
-        if ($errorCode & (Installer::ZIP_OPEN_ERROR | Installer::ZIP_EXTRACT_ERROR)) {
-            return false;
+        if ($tmp) {
+            Folder::delete($tmp);
         }
 
-        return true;
+        return !(is_string($errorCode) || ($errorCode & (Installer::ZIP_OPEN_ERROR | Installer::ZIP_EXTRACT_ERROR)));
+    }
+
+    private static function upgradeGrav($zip, $folder, $keepFolder = false)
+    {
+        static $ignores = [
+            'backup',
+            'cache',
+            'images',
+            'logs',
+            'tmp',
+            'user',
+            '.htaccess',
+            'robots.txt'
+        ];
+
+        if (!is_dir($folder)) {
+            Installer::setError('Invalid source folder');
+        }
+
+        try {
+            $script = $folder . '/system/install.php';
+            /** Install $installer */
+            if ((file_exists($script) && $install = include $script) && is_callable($install)) {
+                $install($zip);
+            } else {
+                Installer::install(
+                    $zip,
+                    GRAV_ROOT,
+                    ['sophisticated' => true, 'overwrite' => true, 'ignore_symlinks' => true, 'ignores' => $ignores],
+                    $folder,
+                    $keepFolder
+                );
+
+                Cache::clearCache();
+            }
+        } catch (\Exception $e) {
+            Installer::setError($e->getMessage());
+        }
     }
 }
